requirements.txt
Django==2.2.20
pytz==2021.1
djangorestframework==3.12.4
